Moved ApplyDefaultSkinSettings
